consolidate some interfaces and classes
