--- src/slskd/Application/IApplication.cs
@@ -1,28 +1,0 @@
-﻿// <copyright file="IApplication.cs" company="slskd Team">
-//     Copyright (c) slskd Team. All rights reserved.
-//
-//     This program is free software: you can redistribute it and/or modify
-//     it under the terms of the GNU Affero General Public License as published
-//     by the Free Software Foundation, either version 3 of the License, or
-//     (at your option) any later version.
-//
-//     This program is distributed in the hope that it will be useful,
-//     but WITHOUT ANY WARRANTY; without even the implied warranty of
-//     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//     GNU Affero General Public License for more details.
-//
-//     You should have received a copy of the GNU Affero General Public License
-//     along with this program.  If not, see https://www.gnu.org/licenses/.
-// </copyright>
-
-namespace slskd
-{
-    using System.Threading.Tasks;
-    using Microsoft.Extensions.Hosting;
-
-    public interface IApplication : IHostedService
-    {
-        public Task CheckVersionAsync();
-        public Task RescanSharesAsync();
-    }
-}
